Shorten the abyssal generator depth to 32 bits (fixes an infinite loop).
A good enough fix would be to re-cast the depth to unsigned during
unmarshalling so it doesn't get sign-extended before putting into an
uint64_t variable.  Bits above 32 can't be ever used, though -- even
bits over 31 hardly ever do (initial value is < 0x80000000 but increases
with time).

+++ crawl-ref/source/abyss.h
@@ -17,7 +17,7 @@ struct abyss_state
 {
     coord_def major_coord;
     uint32_t seed;
-    uint64_t depth;
+    uint32_t depth;
     double phase;
     bool nuke_all;
 };
